Change Basic Iterative Method (BIM)
Fixes #90
