+++ art/attacks/__init__.py
@@ -8,9 +8,9 @@ from art.attacks.carlini import CarliniL2Method, CarliniLInfMethod
 from art.attacks.deepfool import DeepFool
 from art.attacks.elastic_net import ElasticNet
 from art.attacks.fast_gradient import FastGradientMethod
+from art.attacks.projected_gradient_descent import ProjectedGradientDescent
 from art.attacks.iterative_method import BasicIterativeMethod
 from art.attacks.newtonfool import NewtonFool
-from art.attacks.projected_gradient_descent import ProjectedGradientDescent
 from art.attacks.saliency_map import SaliencyMapMethod
 from art.attacks.spatial_transformation import SpatialTransformation
 from art.attacks.universal_perturbation import UniversalPerturbation
